+++ src/features/PluginStore/Loading.tsx
@@ -4,8 +4,8 @@ import { Flexbox } from 'react-layout-kit';
 
 const Loading = memo(() => {
   return (
-    <Flexbox>
-      <Skeleton paragraph={{ rows: 8 }} />
+    <Flexbox padding={16}>
+      <Skeleton active paragraph={{ rows: 8 }} title={false} />
     </Flexbox>
   );
 });